Update index.md
I've deleted the 'key services' and am creating a new sub-section (using the skeleton for 'concepts',) with these. 
The motivation for moving these, was to have some room in order to flesh these out a bit more. 
In that way, the Service Level Agreements document for Corda Network can refer to this section.